# simpeg-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
...
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pegawai_id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  http.ConfirmPasswordRequest:
    properties:
      password:
        type: string
    type: object
...
  http.DeleteAccountRequest:
    properties:
      password:
        type: string
    type: object
  http.EmailVerifyRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  http.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  http.MessageResponse:
    properties:
      message:
        type: string
    type: object
  http.PasswordUpdateRequest:
    properties:
      current_password:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
    type: object
...
  http.ProfileResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      user:
        $ref: '#/definitions/domain.User'
    type: object
  http.ProfileUpdateRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  http.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        type: string
    required:
    - email
    - name
    - password
    type: object
  http.ResetPasswordRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      password_confirmation:
        type: string
      token:
        type: string
    required:
    - email
    - password
    - password_confirmation
    - token
    type: object
...
  http.TokenMessageResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  http.VerifiedResponse:
    properties:
      data:
        properties:
          message:
            type: string
          verified:
            type: boolean
        type: object
    type: object
host: localhost:8000
info:
  contact:
    name: SIMPEG UML
  description: Backend API Sistem Informasi Manajemen Kepegawaian Universitas Muhammadiyah
    Lampung (Gin + GORM).
  termsOfService: http://simpeg.uml.ac.id
  title: SIMPEG UML API
  version: "1.0"
paths:
...
  /auth/account:
    delete:
      consumes:
      - application/json
      description: Menghapus akun user yang sedang login setelah verifikasi password.
      parameters:
      - description: Konfirmasi password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.DeleteAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus akun
      tags:
      - Auth
  /auth/confirm-password:
    post:
      consumes:
      - application/json
      description: Memvalidasi password saat ini (dipakai sebelum aksi sensitif).
      parameters:
      - description: Password saat ini
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.ConfirmPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Konfirmasi password
      tags:
      - Auth
  /auth/email/verify:
    post:
      consumes:
      - application/json
      description: Memverifikasi email menggunakan token dari verify-notification.
      parameters:
      - description: Token verifikasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.EmailVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.VerifiedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Verifikasi email
      tags:
      - Auth
  /auth/email/verify-notification:
    post:
      description: Membuat dan mengirim token verifikasi email untuk user yang sedang
        login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TokenMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kirim ulang link verifikasi email
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Mengirim token reset password ke email. Tanpa SMTP (dev), token
        dikembalikan pada respons.
      parameters:
      - description: Email terdaftar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TokenMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Lupa password
      tags:
      - Auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Mengautentikasi pengguna dan mengembalikan token JWT beserta data
        user.
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login pengguna
      tags:
      - Auth
  /auth/password:
    put:
      consumes:
      - application/json
      description: Mengganti password user yang sedang login setelah memverifikasi
        password saat ini.
      parameters:
      - description: Data password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.PasswordUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ubah password
      tags:
      - Auth
  /auth/profile:
    get:
      description: Mengembalikan data user yang sedang login beserta daftar nama permission.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lihat profil & permission
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Mengubah nama dan email user yang sedang login.
      parameters:
      - description: Data profil
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Perbarui profil
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun baru dan langsung mengembalikan token JWT.
      parameters:
      - description: Data registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Registrasi pengguna baru
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Menetapkan password baru menggunakan token dari forgot-password.
      parameters:
      - description: Data reset
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Reset password
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"